requirements.txt
numpy
flask
redis
pymongo
pymongo[srv]
pillow
face_recognition
tensorflow-cpu